LayoutTests/imported/w3c/web-platform-tests/svg/animations/svglength-additive-by-8-expected.txt


FAIL This tests by-animations adding to previous underlying values assert_approx_equals: expected 30 +/- 1 but got 10
